requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
